api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
sqlalchemy>=2.0
alembic>=1.13
pyjwt>=2.8
slowapi>=0.1.9
python-json-logger>=2.0
anthropic>=0.40.0
httpx>=0.27.0
